+++ main.h
@@ -7,7 +7,7 @@ int print_string(va_list args);
 int print_percent(va_list args);
 int _printf(const char *format, ...);
 int print_int(va_list args);
-int count_digits(int n);
-void print_number(int n);
+int count_digits(long int n);
+void print_number(long int n);
 
 #endif
